# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ANPE CENTRE NORD API",
        "description": "Bienvenue dans la documentation de l'API de l'ANPE CENTRE NORD.<br>Cette API fournit un accès sécurisé aux fonctionnalités du backend, <br>permettant la gestion efficace des ressources liées à l'emploi, aux utilisateurs et aux données associées.",
        "contact": {
            "name": "SYMBOL SARL",
            "email": "admin@admin.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur test de l'API"
        }
    ],
    "paths": {
...
        "/api/secteurs-activite": {
            "get": {
                "tags": [
                    "Secteur d'activite"
                ],
                "summary": "Liste des secteurs d'activité",
                "description": "Retourne la liste des secteurs d'activité",
                "operationId": "8f0a580085512a724fcfbf0b3d552b52",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Secteur d'activite"
                ],
                "summary": "Création d'un secteur d\\'activité",
                "description": "Crée un nouveau secteur d\\'activité et retourne le secteur d\\'activité créé",
                "operationId": "ce2cd430a4a7cde440a6b3c8b5a335a8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "label"
                                ],
                                "properties": {
                                    "label": {
                                        "type": "string",
                                        "example": "Technologie de l'information (IT)"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Lorem ipsum"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "secteur d\\'activité créé avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Les données fournies ne sont pas valides"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/secteurs-activite/{slug}": {
            "get": {
                "tags": [
                    "Secteur d'activite"
                ],
                "summary": "Récupère un secteur d\\'activité par son slug",
                "description": "Retourne un secteur d\\'activité",
                "operationId": "2ef0e341a3f663dbc743628a2a5a1ffe",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "slug du secteur d\\'activité à récupérer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Secteur d'activite"
                ],
                "summary": "Modification d'un secteur d\\'activité",
                "description": "Modifie un secteur d\\'activité et retourne le secteur d\\'activité modifié",
                "operationId": "054e6dedd494d28fe9deb3605007fd20",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "slug du secteur d\\'activité à modifié",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "label"
                                ],
                                "properties": {
                                    "label": {
                                        "type": "string",
                                        "example": "Technologie de l'information (IT)"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Lorem ipsum"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "secteur d\\'activité modifié avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Slug validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "secteur d\\'activité non trouvé"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Les données fournies ne sont pas valides"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Secteur d'activite"
                ],
                "summary": "Supprime un secteur d\\'activité par son slug",
                "description": "Retourne le secteur d\\'activité supprimé",
                "operationId": "de2389cc920fadfdf1fe0cc4e1ef61f4",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "slug du secteur d\\'activité à supprimer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "secteur d\\'activité supprimé avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Slug validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "secteur d\\'activité non trouvé"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}